style: fix clippy issues and format code
