+++ crates/amaru-network/src/tx_submission/tx_submission_client.rs
@@ -12,17 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use crate::tx_submission::tx_client_transport::{PallasTxClientTransport, TxClientTransport};
 use amaru_kernel::peer::Peer;
 use amaru_kernel::{Hash, to_cbor};
 use amaru_ouroboros_traits::{Mempool, MempoolSeqNo, TxId};
-use async_trait::async_trait;
 use minicbor::{CborLen, Encode};
-use pallas_network::miniprotocols::txsubmission;
 use pallas_network::miniprotocols::txsubmission::{EraTxBody, EraTxId, Request, TxIdAndSize};
 use pallas_network::multiplexer::AgentChannel;
 use pallas_traverse::Era;
 use std::collections::VecDeque;
 use std::sync::Arc;
+use tracing::info;
 
 /// Tx submission client state machine for a given peer.
 ///
@@ -65,6 +65,10 @@ impl<Tx: Encode<()> + CborLen<()> + Send + Sync + 'static> TxSubmissionClient<Tx
         mut transport: T,
     ) -> anyhow::Result<()> {
         transport.send_init().await?;
+        info!(peer = %self.peer,
+            "Started tx submission client for peer {}",
+            self.peer
+        );
         loop {
             let request = match transport.next_request().await {
                 Ok(r) => r,
@@ -79,11 +83,11 @@ impl<Tx: Encode<()> + CborLen<()> + Send + Sync + 'static> TxSubmissionClient<Tx
                         transport.send_done().await?;
                         break;
                     }
-                    self.provide_transactions(&mut transport, acknowledged, required_next)
+                    self.send_next_tx_ids(&mut transport, acknowledged, required_next)
                         .await?;
                 }
                 Request::TxIdsNonBlocking(acknowledged, required_next) => {
-                    self.provide_transactions(&mut transport, acknowledged, required_next)
+                    self.send_next_tx_ids(&mut transport, acknowledged, required_next)
                         .await?;
                 }
                 Request::Txs(ids) => {
@@ -109,7 +113,8 @@ impl<Tx: Encode<()> + CborLen<()> + Send + Sync + 'static> TxSubmissionClient<Tx
         Ok(())
     }
 
-    async fn provide_transactions<T: TxClientTransport>(
+    /// Take notice of the acknowledged transactions, and send the next batch of tx ids.
+    async fn send_next_tx_ids<T: TxClientTransport>(
         &mut self,
         transport: &mut T,
         acknowledged: u16,
@@ -155,70 +160,15 @@ impl<Tx: Encode<()> + CborLen<()> + Send + Sync + 'static> TxSubmissionClient<Tx
     }
 }
 
-/// Abstraction over the tx-submission wire used by the client state machine.
-///
-/// This lets us unit-test `TxSubmissionClient` without needing a real
-/// `AgentChannel` / `TcpStream`. Production code uses the pallas
-/// `txsubmission::Client<AgentChannel>` through the adapter below.
-#[async_trait]
-pub trait TxClientTransport: Send {
-    async fn send_init(&mut self) -> anyhow::Result<()>;
-    async fn send_done(&mut self) -> anyhow::Result<()>;
-
-    async fn next_request(&mut self) -> anyhow::Result<Request<EraTxId>>;
-
-    async fn reply_tx_ids(&mut self, ids: Vec<TxIdAndSize<EraTxId>>) -> anyhow::Result<()>;
-    async fn reply_txs(&mut self, txs: Vec<EraTxBody>) -> anyhow::Result<()>;
-}
-
-/// Production adapter around pallas' txsubmission client.
-pub struct PallasTxClientTransport {
-    inner: txsubmission::Client,
-}
-
-impl PallasTxClientTransport {
-    pub fn new(agent_channel: AgentChannel) -> Self {
-        Self {
-            inner: txsubmission::Client::new(agent_channel),
-        }
-    }
-}
-
-#[async_trait]
-impl TxClientTransport for PallasTxClientTransport {
-    async fn send_init(&mut self) -> anyhow::Result<()> {
-        Ok(self.inner.send_init().await?)
-    }
-
-    async fn send_done(&mut self) -> anyhow::Result<()> {
-        Ok(self.inner.send_done().await?)
-    }
-
-    async fn next_request(&mut self) -> anyhow::Result<Request<EraTxId>> {
-        Ok(self.inner.next_request().await?)
-    }
-
-    async fn reply_tx_ids(&mut self, ids: Vec<TxIdAndSize<EraTxId>>) -> anyhow::Result<()> {
-        Ok(self.inner.reply_tx_ids(ids).await?)
-    }
-
-    async fn reply_txs(&mut self, txs: Vec<EraTxBody>) -> anyhow::Result<()> {
-        Ok(self.inner.reply_txs(txs).await?)
-    }
-}
-
 #[cfg(test)]
 pub(crate) mod tests {
     use super::*;
-    use amaru_kernel::cbor::Encode;
+    use crate::tx_submission::tests::{
+        Tx, assert_next_message, assert_tx_bodies_reply, assert_tx_ids_reply,
+    };
+    use crate::tx_submission::tx_client_transport::tests::MockClientTransport;
     use amaru_mempool::strategies::InMemoryMempool;
-    use minicbor::{Decode, Decoder, Encoder};
-    use minicbor::encode::{Error, Write};
     use pallas_network::miniprotocols::txsubmission::Message;
-    use pallas_network::miniprotocols::txsubmission::Message::{ReplyTxIds, ReplyTxs};
-    use std::fmt::{Debug, Display};
-
-    use crate::tx_submission::tx_submission_server::EraTxIdOrd;
     use tokio::sync::mpsc;
     use tokio::sync::mpsc::{Receiver, Sender};
     use tokio::task::JoinHandle;
@@ -279,7 +229,7 @@ pub(crate) mod tests {
     async fn serve_transactions_non_blocking() -> anyhow::Result<()> {
         // Create a mempool with some transactions
         let mempool: Arc<InMemoryMempool<Tx>> = Arc::new(InMemoryMempool::default());
-        let txs = vec![
+        let txs = [
             Tx::new("0d8d00cdd4657ac84d82f0a56067634a"),
             Tx::new("1d8d00cdd4657ac84d82f0a56067634a"),
             Tx::new("2d8d00cdd4657ac84d82f0a56067634a"),
@@ -342,49 +292,6 @@ pub(crate) mod tests {
 
     // HELPERS
 
-    /// Check that the next message is a ReplyTxIds with the expected ids.
-    pub(crate) async fn assert_tx_ids_reply(
-        rx_messages: &mut Receiver<Message<EraTxId, EraTxBody>>,
-        era_tx_ids: &[EraTxId],
-        expected_ids: &[usize],
-    ) -> anyhow::Result<()> {
-        let tx_ids_and_sizes: Vec<TxIdAndSize<EraTxId>> = expected_ids
-            .iter()
-            .map(|&i| TxIdAndSize(era_tx_ids[i].clone(), 32))
-            .collect();
-        assert_next_message(rx_messages, ReplyTxIds(tx_ids_and_sizes)).await?;
-        Ok(())
-    }
-
-    /// Check that the next message is a ReplyTxs with the expected transaction bodies.
-    pub(crate) async fn assert_tx_bodies_reply(
-        rx_messages: &mut Receiver<Message<EraTxId, EraTxBody>>,
-        era_tx_bodies: &[EraTxBody],
-        expected_body_ids: &[usize],
-    ) -> anyhow::Result<()> {
-        let txs: Vec<EraTxBody> = expected_body_ids
-            .iter()
-            .map(|&i| era_tx_bodies[i].clone())
-            .collect();
-        assert_next_message(rx_messages, ReplyTxs(txs)).await?;
-        Ok(())
-    }
-
-    /// Check that the next message matches the expected one.
-    pub(crate) async fn assert_next_message(
-        rx_messages: &mut Receiver<Message<EraTxId, EraTxBody>>,
-        expected: Message<EraTxId, EraTxBody>,
-    ) -> anyhow::Result<()> {
-        let actual = rx_messages
-            .recv()
-            .await
-            .ok_or_else(|| anyhow::anyhow!("no message received"))?;
-        let actual = MessageEq(actual);
-        let expected = MessageEq(expected);
-        assert_eq!(actual, expected, "actual = {actual}\nexpected = {expected}");
-        Ok(())
-    }
-
     /// Send a series of requests to a tx submission client backed by the given mempool.
     /// Returns a receiver to collect the messages that the client sent back.
     async fn start_client(
@@ -394,18 +301,15 @@ pub(crate) mod tests {
         Receiver<Message<EraTxId, EraTxBody>>,
         JoinHandle<anyhow::Result<()>>,
     )> {
-        let mut client_sm = TxSubmissionClient::new(mempool.clone(), Peer::new("peer-1"));
+        let mut client = TxSubmissionClient::new(mempool.clone(), Peer::new("peer-1"));
 
         let (tx_req, rx_req) = mpsc::channel(10);
         let (tx_messages, rx_messages) = mpsc::channel(10);
 
-        let transport = MockClientTransport {
-            rx_req,
-            tx_reply: tx_messages,
-        };
+        let transport = MockClientTransport::new(rx_req, tx_messages);
 
         let client_handle =
-            tokio::spawn(async move { client_sm.start_client_with_transport(transport).await });
+            tokio::spawn(async move { client.start_client_with_transport(transport).await });
         Ok((tx_req, rx_messages, client_handle))
     }
 
@@ -417,168 +321,5 @@ pub(crate) mod tests {
     /// Create a new EraTxBody for the Conway era.
     pub(crate) fn new_era_tx_body(tx_body: Vec<u8>) -> EraTxBody {
         EraTxBody(Era::Conway.into(), tx_body)
-    }
-
-    struct MockClientTransport {
-        // server -> client messages
-        rx_req: Receiver<Request<EraTxId>>,
-        // client -> server messages
-        tx_reply: Sender<Message<EraTxId, EraTxBody>>,
-    }
-
-    fn era_tx_id_to_string(era_tx_id: &EraTxId) -> String {
-        Hash::<32>::from(era_tx_id.1.as_slice()).to_string()
-    }
-
-    fn era_tx_body_to_string(era_tx_body: &EraTxBody) -> String {
-        String::from_utf8(era_tx_body.1.clone()).unwrap()
-    }
-
-    #[async_trait::async_trait]
-    impl TxClientTransport for MockClientTransport {
-        async fn send_init(&mut self) -> anyhow::Result<()> {
-            self.tx_reply.send(Message::Init).await?;
-            Ok(())
-        }
-
-        async fn send_done(&mut self) -> anyhow::Result<()> {
-            self.tx_reply.send(Message::Done).await?;
-            Ok(())
-        }
-
-        async fn next_request(&mut self) -> anyhow::Result<Request<EraTxId>> {
-            self.rx_req
-                .recv()
-                .await
-                .ok_or_else(|| anyhow::anyhow!("mock closed"))
-        }
-
-        async fn reply_tx_ids(&mut self, ids: Vec<TxIdAndSize<EraTxId>>) -> anyhow::Result<()> {
-            self.tx_reply.send(Message::ReplyTxIds(ids)).await?;
-            Ok(())
-        }
-
-        async fn reply_txs(&mut self, txs: Vec<EraTxBody>) -> anyhow::Result<()> {
-            self.tx_reply.send(Message::ReplyTxs(txs)).await?;
-            Ok(())
-        }
-    }
-
-    #[derive(Debug, PartialEq, Eq, Clone)]
-    pub(crate) struct Tx {
-        tx_body: String,
-    }
-
-    impl Tx {
-        pub(crate) fn new(tx_body: impl Into<String>) -> Self {
-            Self {
-                tx_body: tx_body.into(),
-            }
-        }
-
-        pub(crate) fn tx_id(&self) -> TxId {
-            TxId::from(self.tx_body.as_str())
-        }
-
-        pub(crate) fn tx_body(&self) -> Vec<u8> {
-            minicbor::to_vec(&self.tx_body).unwrap()
-        }
-    }
-
-    impl CborLen<()> for Tx {
-        fn cbor_len(&self, _ctx: &mut ()) -> usize {
-            self.tx_body.len()
-        }
-    }
-
-    impl Encode<()> for Tx {
-        fn encode<W: Write>(
-            &self,
-            e: &mut Encoder<W>,
-            _ctx: &mut (),
-        ) -> Result<(), Error<W::Error>> {
-            e.encode(&self.tx_body)?;
-            Ok(())
-        }
-    }
-
-    impl<'a> Decode<'a, ()> for Tx {
-        fn decode(d: &mut Decoder<'a>, _ctx: &mut ()) -> Result<Self, minicbor::decode::Error> {
-            let tx_body: String = d.decode()?;
-            Ok(Tx { tx_body })
-        }
-    }
-
-    pub(crate) struct MessageEq(Message<EraTxId, EraTxBody>);
-
-    impl Debug for MessageEq {
-        fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-            write!(f, "{}", self)
-        }
-    }
-
-    impl Display for MessageEq {
-        fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-            match &self.0 {
-                Message::Init => write!(f, "Init"),
-                Message::RequestTxIds(blocking, ack, req) => {
-                    write!(
-                        f,
-                        "RequestTxIds(blocking={}, ack={}, req={})",
-                        blocking, ack, req
-                    )
-                }
-                ReplyTxIds(ids_and_sizes) => {
-                    let ids: Vec<String> = ids_and_sizes
-                        .iter()
-                        .map(|tx_id_and_size| era_tx_id_to_string(&tx_id_and_size.0))
-                        .collect();
-                    write!(f, "ReplyTxIds(ids=[{}])", ids.join(", "))
-                }
-                Message::RequestTxs(ids) => {
-                    let ids_str: Vec<String> =
-                        ids.iter().map(|tx_id| era_tx_id_to_string(tx_id)).collect();
-                    write!(f, "RequestTxs(ids=[{}])", ids_str.join(", "))
-                }
-                ReplyTxs(bodies) => {
-                    let bodies_str: Vec<String> = bodies
-                        .iter()
-                        .map(|tx_body| era_tx_body_to_string(tx_body))
-                        .collect();
-                    write!(f, "ReplyTxs(bodies=[{}])", bodies_str.join(", "))
-                }
-                Message::Done => write!(f, "Done"),
-            }
-        }
-    }
-
-    impl PartialEq for MessageEq {
-        fn eq(&self, other: &Self) -> bool {
-            match (&self.0, &other.0) {
-                (Message::Done, Message::Done) => true,
-                (Message::Init, Message::Init) => true,
-                (Message::RequestTxIds(b1, a1, r1), Message::RequestTxIds(b2, a2, r2)) => {
-                    b1 == b2 && a1 == a2 && r1 == r2
-                }
-                (Message::RequestTxs(ids1), Message::RequestTxs(ids2)) => {
-                    ids1.into_iter()
-                        .map(|id| EraTxIdOrd::new(id.clone()))
-                        .collect::<Vec<_>>()
-                        == ids2
-                        .into_iter()
-                        .map(|id| EraTxIdOrd::new(id.clone()))
-                        .collect::<Vec<_>>()
-                }
-                (ReplyTxIds(ids1), ReplyTxIds(ids2)) => {
-                    ids1.into_iter().map(|id| (id.0.0, id.0.1.clone(), id.1)).collect::<Vec<_>>()
-                        == ids2.into_iter().map(|id| (id.0.0, id.0.1.clone(), id.1)).collect::<Vec<_>>()
-                }
-                (ReplyTxs(txs1), ReplyTxs(txs2)) => {
-                    txs1.into_iter().cloned().map(|tx| (tx.0, tx.1)).collect::<Vec<_>>()
-                        == txs2.into_iter().cloned().map(|tx| (tx.0, tx.1)).collect::<Vec<_>>()
-                }
-                _ => false,
-            }
-        }
     }
 }
